+++ packages/venia-ui/lib/components/ProductFullDetail/productFullDetail.js
@@ -1,8 +1,7 @@
-import React, { Suspense, useCallback, useState } from 'react';
+import React, { Suspense } from 'react';
 import { arrayOf, bool, number, shape, string } from 'prop-types';
 import { Form } from 'informed';
 import { Price } from '@magento/peregrine';
-import { useCartContext } from '@magento/peregrine/lib/context/cart';
 import defaultClasses from './productFullDetail.css';
 import { mergeClasses } from '../../classify';
 
@@ -12,143 +11,42 @@ import Carousel from '../ProductImageCarousel';
 import Quantity from '../ProductQuantity';
 import RichText from '../RichText';
 
-import appendOptionsToPayload from '../../util/appendOptionsToPayload';
-import findMatchingVariant from '../../util/findMatchingProductVariant';
-import isProductConfigurable from '../../util/isProductConfigurable';
+import { useProductFullDetail } from '@magento/peregrine/lib/talons/ProductFullDetail/useProductFullDetail';
 
 const Options = React.lazy(() => import('../ProductOptions'));
 
-const INITIAL_OPTION_CODES = new Map();
-const INITIAL_OPTION_SELECTIONS = new Map();
-const INITIAL_QUANTITY = 1;
-
-const deriveOptionCodesFromProduct = product => {
-    // If this is a simple product it has no option codes.
-    if (!isProductConfigurable(product)) {
-        return INITIAL_OPTION_CODES;
-    }
-
-    // Initialize optionCodes based on the options of the product.
-    const initialOptionCodes = new Map();
-    for (const {
-        attribute_id,
-        attribute_code
-    } of product.configurable_options) {
-        initialOptionCodes.set(attribute_id, attribute_code);
-    }
-
-    return initialOptionCodes;
-};
-
-const getIsMissingOptions = (product, optionSelections) => {
-    // Non-configurable products can't be missing options.
-    if (!isProductConfigurable(product)) {
-        return false;
-    }
-
-    // Configurable products are missing options if we have fewer
-    // option selections than the product has options.
-    const { configurable_options } = product;
-    const numProductOptions = configurable_options.length;
-    const numProductSelections = optionSelections.size;
-
-    return numProductSelections < numProductOptions;
-};
-
-const getMediaGalleryEntries = (product, optionCodes, optionSelections) => {
-    let value = [];
-
-    const { media_gallery_entries, variants } = product;
-    const isConfigurable = isProductConfigurable(product);
-    const optionsSelected = optionSelections.size > 0;
-
-    if (!isConfigurable || !optionsSelected) {
-        value = media_gallery_entries;
-    } else {
-        const item = findMatchingVariant({
-            optionCodes,
-            optionSelections,
-            variants
-        });
-
-        value = item
-            ? [...item.product.media_gallery_entries, ...media_gallery_entries]
-            : media_gallery_entries;
-    }
-
-    const key = value.reduce((fullKey, entry) => {
-        return `${fullKey},${entry.file}`;
-    }, '');
-
-    return { key, value };
-};
-
 const ProductFullDetail = props => {
-    const [{ isAddingItem }, { addItemToCart }] = useCartContext();
-
-    // Props.
     const { product } = props;
 
-    // State.
-    const [quantity, setQuantity] = useState(INITIAL_QUANTITY);
-    const [optionSelections, setOptionSelections] = useState(
-        INITIAL_OPTION_SELECTIONS
-    );
-    const derivedOptionCodes = deriveOptionCodesFromProduct(product);
-    const [optionCodes] = useState(derivedOptionCodes);
+    const talonProps = useProductFullDetail({
+        product
+    });
 
-    // Members.
-    const { amount: productPrice } = product.price.regularPrice;
+    const {
+        handleAddToCart,
+        handleSelectionChange,
+        handleSetQuantity,
+        isAddToCartDisabled,
+        mediaGalleryEntries,
+        productDetails,
+        quantity
+    } = talonProps;
+
     const classes = mergeClasses(defaultClasses, props.classes);
-    const isMissingOptions = getIsMissingOptions(product, optionSelections);
-    const mediaGalleryEntries = getMediaGalleryEntries(
-        product,
-        optionCodes,
-        optionSelections
-    );
-
-    // Event handlers.
-    const handleAddToCart = useCallback(() => {
-        const payload = {
-            item: product,
-            productType: product.__typename,
-            quantity
-        };
-
-        if (isProductConfigurable(product)) {
-            appendOptionsToPayload(payload, optionSelections, optionCodes);
-        }
-
-        addItemToCart(payload);
-    }, [addItemToCart, optionCodes, optionSelections, product, quantity]);
-
-    const handleSelectionChange = useCallback(
-        (optionId, selection) => {
-            // We must create a new Map here so that React knows that the value
-            // of optionSelections has changed.
-            const newOptionSelections = new Map([...optionSelections]);
-            newOptionSelections.set(optionId, Array.from(selection).pop());
-            setOptionSelections(newOptionSelections);
-        },
-        [optionSelections]
-    );
 
     return (
         <Form className={classes.root}>
             <section className={classes.title}>
-                <h1 className={classes.productName}>{product.name}</h1>
+                <h1 className={classes.productName}>{productDetails.name}</h1>
                 <p className={classes.productPrice}>
                     <Price
-                        currencyCode={productPrice.currency}
-                        value={productPrice.value}
+                        currencyCode={productDetails.price.currency}
+                        value={productDetails.price.value}
                     />
                 </p>
             </section>
             <section className={classes.imageCarousel}>
-                <Carousel
-                    images={mediaGalleryEntries.value}
-                    key={mediaGalleryEntries.key}
-                />
+                <Carousel images={mediaGalleryEntries} />
             </section>
             <section className={classes.options}>
                 <Suspense fallback={fullPageLoadingIndicator}>
@@ -160,13 +58,16 @@ const ProductFullDetail = props => {
             </section>
             <section className={classes.quantity}>
                 <h2 className={classes.quantityTitle}>Quantity</h2>
-                <Quantity initialValue={quantity} onValueChange={setQuantity} />
+                <Quantity
+                    initialValue={quantity}
+                    onValueChange={handleSetQuantity}
+                />
             </section>
             <section className={classes.cartActions}>
                 <Button
                     priority="high"
                     onClick={handleAddToCart}
-                    disabled={isAddingItem || isMissingOptions}
+                    disabled={isAddToCartDisabled}
                 >
                     Add to Cart
                 </Button>
@@ -175,11 +76,11 @@ const ProductFullDetail = props => {
                 <h2 className={classes.descriptionTitle}>
                     Product Description
                 </h2>
-                <RichText content={product.description} />
+                <RichText content={productDetails.description} />
             </section>
             <section className={classes.details}>
                 <h2 className={classes.detailsTitle}>SKU</h2>
-                <strong>{product.sku}</strong>
+                <strong>{productDetails.sku}</strong>
             </section>
         </Form>
     );
